requirements.txt
pyserial >= 3.5.0
pyinstaller >= 4.10